pyproject.toml
[project]
name = "comunitu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.0",
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "stripe>=12.2.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "ipdb>=0.13.13",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
]

[tool.black]
exclude = '''
(
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | migrations
)
'''
